requirements.txt
pandas
yfinance
prophet
google-genai
matplotlib
feedparser
python-dotenv
streamlit
plotly
